APITASK/requirements.txt
Django==3.0
djangorestframework==3.11.1
django-environ==0.4.5
psycopg2-binary>=2.8
djangorestframework-xml==2.0.0